+++ com.unity.cinemachine/Samples~/3D Samples/Clearshot/RunnerController.cs
@@ -6,23 +6,31 @@ namespace Cinemachine.Examples
     [RequireComponent(typeof(CinemachineSplineCart))]
     public class RunnerController : MonoBehaviour
     {
+        public float WaitTimeAtStart = 1;
+
         CinemachineSplineCart m_Cart;
-
-        void Start() => m_Cart = GetComponent<CinemachineSplineCart>();
-
         static bool s_LeaderWasSlowed;
         bool m_IsTired;
+        float m_StartTime;
+
+        void OnEnable()
+        {
+            m_Cart = GetComponent<CinemachineSplineCart>();
+            ResetRace();
+        }
 
         void Update()
         {
-            // Slow down leader to improve chances of at least 1 take over in one run
+            m_Cart.AutomaticDolly.Enabled = Time.time > m_StartTime + WaitTimeAtStart;
+
+            // Slow down leader to improve chances of at least 1 takeover per run
             if (!s_LeaderWasSlowed)
             {
                 if (m_Cart.Spline.Spline.ConvertIndexUnit(
                     m_Cart.SplinePosition, m_Cart.PositionUnits, PathIndexUnit.Normalized) > 0.5f)
                 {
                     s_LeaderWasSlowed = true;
-                    if (m_Cart.AutomaticDolly.Implementation is RandomizedDollySpeed speedControl)
+                    if (m_Cart.AutomaticDolly.Method is RandomizedDollySpeed speedControl)
                     {
                         // Leader is tired!
                         speedControl.MinSpeed /= 2;
@@ -36,10 +44,15 @@ namespace Cinemachine.Examples
         // This is called by the "Restart Race" UX button.
         public void ResetRace()
         {
+            // Reset position to start
             m_Cart.SplinePosition = 0;
-            if (m_Cart.AutomaticDolly.Implementation is RandomizedDollySpeed speedControl)
+            m_StartTime = Time.time;
+            m_Cart.AutomaticDolly.Enabled = false;
+
+            // Reset slowdown mechanism
+            if (m_Cart.AutomaticDolly.Method is RandomizedDollySpeed speedControl)
             {
-                m_Cart.AutomaticDolly.Implementation.Reset();
+                m_Cart.AutomaticDolly.Method.Reset();
                 if (m_IsTired)
                 {
                     // Restore the speed, leader is no longer tired